requirements.txt
catt
adafruit_ssd1306
retrying
Pillow